Refeactor to preprare for integrating cast store
