+++ src/core/mod.rs
@@ -1,1 +1,3 @@
+pub mod error;
 pub mod types;
+pub mod util;
