+++ src/storage/mod.rs
@@ -1,5 +1,5 @@
+mod constants;
 pub mod db;
-pub mod hub_error;
 pub mod store;
 pub mod trie;
 pub mod util;
