+++ src/storage/store/account/message.rs
@@ -1,11 +1,13 @@
 use prost::Message as _;
 
+use super::PAGE_SIZE_MAX;
+use crate::core::error::HubError;
+use crate::storage::constants::{RootPrefix, UserPostfix};
+use crate::storage::db::{PageOptions, RocksdbError};
 use crate::{
-    db::{RocksDB, RocksDbTransactionBatch},
-    protos::{CastId, Message as MessageProto, MessageData, MessageType},
+    proto::message::{CastId, Message as MessageProto, MessageData, MessageType},
+    storage::db::{RocksDB, RocksDbTransactionBatch},
 };
-
-use super::{store::HubError, PageOptions, PAGE_SIZE_MAX};
 
 pub const FID_BYTES: usize = 4;
 
@@ -13,124 +15,6 @@ pub const TS_HASH_LENGTH: usize = 24;
 pub const HASH_LENGTH: usize = 20;
 
 pub const TRUE_VALUE: u8 = 1;
-
-/** Copied from the JS code */
-#[allow(dead_code)]
-pub enum RootPrefix {
-    /* Used for multiple purposes, starts with a 4-byte fid */
-    User = 1,
-    /* Used to index casts by parent */
-    CastsByParent = 2,
-    /* Used to index casts by mention */
-    CastsByMention = 3,
-    /* Used to index links by target */
-    LinksByTarget = 4,
-    /* Used to index reactions by target  */
-    ReactionsByTarget = 5,
-    /* Deprecated */
-    // IdRegistryEvent = 6,
-    // NameRegistryEvent = 7,
-    // IdRegistryEventByCustodyAddress = 8,
-    /* Used to store the state of the hub */
-    HubState = 9,
-    /* Revoke signer jobs */
-    JobRevokeMessageBySigner = 10,
-    /* Sync Merkle Trie Node */
-    SyncMerkleTrieNode = 11,
-    /* Deprecated */
-    // JobUpdateNameExpiry = 12,
-    // NameRegistryEventsByExpiry = 13,
-    /* To check if the Hub was cleanly shutdown */
-    HubCleanShutdown = 14,
-    /* Event log */
-    HubEvents = 15,
-    /* The network ID that the rocksDB was created with */
-    Network = 16,
-    /* Used to store fname server name proofs */
-    FNameUserNameProof = 17,
-    /* Used to store gossip network metrics */
-    // Deprecated, DO NOT USE
-    // GossipMetrics = 18,
-
-    /* Used to index user submitted username proofs */
-    UserNameProofByName = 19,
-
-    // Deprecated
-    // RentRegistryEvent = 20,
-    // RentRegistryEventsByExpiry = 21,
-    // StorageAdminRegistryEvent = 22,
-
-    /* Used to store on chain events */
-    OnChainEvent = 23,
-
-    /** DB Schema version */
-    DBSchemaVersion = 24,
-
-    /* Used to index verifications by address */
-    VerificationByAddress = 25,
-
-    /* Store the connected peers */
-    ConnectedPeers = 26,
-
-    /* Used to index fname username proofs by fid */
-    FNameUserNameProofByFid = 27,
-}
-
-/** Copied from the JS code */
-#[repr(u8)]
-pub enum UserPostfix {
-    /* Message records (1-85) */
-    CastMessage = 1,
-    LinkMessage = 2,
-    ReactionMessage = 3,
-    VerificationMessage = 4,
-    // Deprecated
-    // SignerMessage = 5,
-    UserDataMessage = 6,
-    UsernameProofMessage = 7,
-
-    // Add new message types here
-    // NOTE: If you add a new message type, make sure that it is only used to store Message protobufs.
-    // If you need to store an index, use one of the UserPostfix values below (>86).
-    /** Index records (must be 86-255) */
-    // Deprecated
-    // BySigner = 86, // Index message by its signer
-
-    /** CastStore add and remove sets */
-    CastAdds = 87,
-    CastRemoves = 88,
-
-    /* LinkStore add and remove sets */
-    LinkAdds = 89,
-    LinkRemoves = 90,
-
-    /** ReactionStore add and remove sets */
-    ReactionAdds = 91,
-    ReactionRemoves = 92,
-
-    /** Verification add and remove sets */
-    VerificationAdds = 93,
-    VerificationRemoves = 94,
-
-    /* Deprecated */
-    // SignerAdds = 95,
-    // SignerRemoves = 96,
-
-    /* UserDataStore add set */
-    UserDataAdds = 97,
-
-    /* UserNameProof add set */
-    UserNameProofAdds = 99,
-
-    /* Link Compact State set */
-    LinkCompactStateMessage = 100,
-}
-
-impl UserPostfix {
-    pub fn as_u8(self) -> u8 {
-        self as u8
-    }
-}
 
 /** A page of messages returned from various APIs */
 pub struct MessagesPage {
@@ -273,10 +157,7 @@ pub fn get_message(
     match db.get(&key)? {
         Some(bytes) => match message_decode(bytes.as_slice()) {
             Ok(message) => Ok(Some(message)),
-            Err(_) => Err(HubError {
-                code: "db.internal_error".to_string(),
-                message: "could not decode message".to_string(),
-            }),
+            Err(e) => Err(e.into()),
         },
         None => Ok(None),
     }
@@ -317,7 +198,7 @@ where
     let mut messages_bytes = Vec::new();
     let mut last_key = vec![];
 
-    db.for_each_iterator_by_prefix(prefix, page_options, |key, value| {
+    db.for_each_iterator_by_prefix(Some(prefix.to_vec()), None, page_options, |key, value| {
         match message_decode(value) {
             Ok(message) => {
                 if filter(&message) {
@@ -331,10 +212,7 @@ where
 
                 Ok(false) // Continue iterating
             }
-            Err(e) => Err(HubError {
-                code: "db.internal_error".to_string(),
-                message: format!("could not decode message: {}", e),
-            }),
+            Err(e) => Err(HubError::from(e)),
         }
     })?;
 
@@ -362,7 +240,7 @@ pub fn message_encode(message: &MessageProto) -> Vec<u8> {
     }
 }
 
-pub fn message_decode(bytes: &[u8]) -> Result<MessageProto, HubError> {
+pub fn message_decode(bytes: &[u8]) -> Result<MessageProto, RocksdbError> {
     if let Ok(mut msg) = MessageProto::decode(bytes) {
         if msg.data.is_none()
             && msg.data_bytes.is_some()
@@ -375,10 +253,7 @@ pub fn message_decode(bytes: &[u8]) -> Result<MessageProto, HubError> {
 
         Ok(msg)
     } else {
-        Err(HubError {
-            code: "db.internal_error".to_string(),
-            message: "could not decode message".to_string(),
-        })
+        Err(RocksdbError::DecodeError)
     }
 }
