+++ src/storage/store/mod.rs
@@ -1,6 +1,7 @@
 pub use self::block::*;
 pub use self::utils::*;
 
+pub mod account;
 pub mod block;
 pub mod engine;
 pub mod shard;
