+++ src/storage/store/shard.rs
@@ -1,11 +1,11 @@
+use super::super::constants::PAGE_SIZE_MAX;
+use crate::core::error::HubError;
 use crate::proto::snapchain::{Block, ShardChunk, Transaction};
+use crate::storage::constants::RootPrefix;
 use crate::storage::db::{PageOptions, RocksDB, RocksdbError};
-use crate::storage::store::block::RootPrefix;
 use prost::Message;
 use std::sync::Arc;
 use thiserror::Error;
-
-use super::utils::PAGE_SIZE_MAX;
 
 // TODO(aditi): This code definitely needs unit tests
 #[derive(Error, Debug)]
@@ -49,7 +49,7 @@ fn get_shard_page_by_prefix(
     let mut last_key = vec![];
 
     db.for_each_iterator_by_prefix_paged(start_prefix, stop_prefix, page_options, |key, value| {
-        let block = ShardChunk::decode(value)?;
+        let block = ShardChunk::decode(value).map_err(|e| HubError::from(e))?;
         shard_chunks.push(block);
 
         if shard_chunks.len() >= page_options.page_size.unwrap_or(PAGE_SIZE_MAX) {
@@ -58,7 +58,8 @@ fn get_shard_page_by_prefix(
         }
 
         Ok(false) // Continue iterating
-    })?;
+    })
+    .map_err(|e| ShardStorageError::TooManyShardsInResult)?; // TODO: Return the right error
 
     let next_page_token = if last_key.len() > 0 {
         Some(last_key)
